[Guardian] modularize hashi-types/guardian; rename CommitteeTransition → CommitteeTransitionRequest
Break the two grab-bag files in hashi-types/guardian into focused modules.
Pure moves: all public paths are preserved via re-exports, so external
consumers (monitor, hashi, tools) are unchanged apart from the rename below.

- signing.rs: GuardianSigned envelope + IntentType/SigningIntent, with the
  struct and its new/verify impl co-located (impl moved out of crypto.rs).
- log/{mod,message,envelope}.rs: the LogMessage family (message.rs) and the
  LogRecord wrapper + object-key/lock-duration logic (envelope.rs); the S3
  layout constants live in log/mod.rs.
- bitcoin_utils/{mod,utxo,taproot}.rs: UTXO/transaction types + signing (utxo)
  vs the 2-of-2 taproot descriptor/address/derivation (taproot); BTC_LIB and
  the DerivationPath alias stay in mod.rs.

mod.rs drops from ~1266 to ~640 lines. The request/response domain types are
intentionally left there.

Minor consolidations (no behavior change): the two impl WithdrawalLogMessage
blocks merged into one; object_key tests relocated into log/envelope.rs; the
one bitcoin_utils integration test now builds its outputs via the public
OutputUTXO::new_external/new_internal constructors (private fields aren't
reachable from the test's new module).

Also folds in the CommitteeTransition → CommitteeTransitionRequest rename
across the guardian module and its consumers (hashi-guardian, hashi), plus a
guardian_service.proto comment tweak.

Deferred to a follow-up: standardizing proto_conversions on From/TryFrom.
