+++ crates/hashi-types/src/guardian/mod.rs
@@ -4,7 +4,9 @@
 pub mod bitcoin_utils;
 pub mod crypto;
 pub mod errors;
+pub mod log;
 pub mod proto_conversions;
+pub mod signing;
 pub mod test_utils;
 pub mod time_utils;
 
@@ -14,6 +16,10 @@ pub mod s3_utils;
 pub use limiter::LimiterConfig;
 pub use limiter::LimiterState;
 pub use limiter::RateLimiter;
+pub use log::*;
+pub use signing::GuardianSigned;
+pub use signing::IntentType;
+pub use signing::SigningIntent;
 pub use time_utils::UnixMillis;
 pub use time_utils::now_timestamp_ms;
 pub use time_utils::now_timestamp_secs;
@@ -26,9 +32,7 @@ use self::bitcoin_utils::TxUTXOsWire;
 use self::errors::GuardianError::*;
 pub use crate::committee::Committee as HashiCommittee;
 pub use crate::committee::CommitteeMember as HashiCommitteeMember;
-use crate::committee::CommitteeSignature;
 pub use crate::committee::SignedMessage as HashiSigned;
-use crate::guardian::s3_utils::S3HourScopedDirectory;
 use crate::pgp::PgpPublicCert;
 pub use bitcoin::Address as BitcoinAddress;
 pub use bitcoin::secp256k1::Keypair as BitcoinKeypair;
@@ -44,40 +48,14 @@ pub use ed25519_consensus::Signature as GuardianSignature;
 pub use ed25519_consensus::SigningKey as GuardianSignKeyPair;
 pub use ed25519_consensus::VerificationKey as GuardianPubKey;
 pub use errors::*;
-/// The raw MPC verifying key as a curve point — y-parity preserved.
-/// Used wherever the 2-of-2 descriptor needs to derive a child key that
-/// matches the MPC's signing protocol, which works directly on the raw `G`.
 pub use fastcrypto_tbls::threshold_schnorr::G as HashiMasterG;
 use rand_core::CryptoRng;
 use rand_core::RngCore;
 use serde::Deserialize;
 use serde::Serialize;
-use std::time::Duration;
 // ---------------------------------
 //          Constants
 // ---------------------------------
-
-/// Object lock durations used for S3 log objects.
-///
-/// These are public so that external verifiers/monitors can apply the same expectations.
-///
-/// TODO: Uniform 7 days is a coarse placeholder. Revisit per-log-type
-/// against the SLO we actually want to defend — heartbeats could be shorter,
-/// ceremony/withdraws likely want years.
-const ONE_WEEK: Duration = Duration::from_secs(7 * 24 * 60 * 60);
-pub const S3_OBJECT_LOCK_DURATION_INIT: Duration = ONE_WEEK;
-pub const S3_OBJECT_LOCK_DURATION_WITHDRAW: Duration = ONE_WEEK;
-pub const S3_OBJECT_LOCK_DURATION_HEARTBEAT: Duration = ONE_WEEK;
-pub const S3_OBJECT_LOCK_DURATION_CEREMONY: Duration = ONE_WEEK;
-pub const S3_OBJECT_LOCK_DURATION_COMMITTEE_UPDATE: Duration = ONE_WEEK;
-
-/// S3 sub-prefixes used for guardian log streams.
-/// See `crates/hashi-guardian/README.md` for canonical key layout.
-pub const S3_DIR_INIT: &str = "init";
-pub const S3_DIR_WITHDRAW: &str = "withdraw";
-pub const S3_DIR_HEARTBEAT: &str = "heartbeat";
-pub const S3_DIR_CEREMONY: &str = "ceremony";
-pub const S3_DIR_COMMITTEE_UPDATE: &str = "committee-update";
 
 /// Length of the session ID prefix (hex chars) used in S3 keys. 16 hex =
 /// 64 bits of the signing pubkey, comfortably below any collision risk for
@@ -103,78 +81,8 @@ pub enum EnclaveMode {
 }
 
 // ---------------------------------
-//          Intents
+//    Common requests and responses
 // ---------------------------------
-
-/// All possible signing intent types.
-/// Using an enum ensures no two types can accidentally share the same intent value.
-#[repr(u8)]
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-pub enum IntentType {
-    /// Intent for all LogMessage's
-    LogMessage = 0,
-    /// Intent for SetupNewKeyResponse
-    SetupNewKeyResponse = 1,
-    /// Intent for StandardWithdrawalResponse
-    StandardWithdrawalResponse = 2,
-    /// Intent for GuardianInfo
-    GuardianInfo = 3,
-    /// Intent for RotateKpsResponse
-    RotateKpsResponse = 4,
-}
-
-/// Trait for types that can be signed, providing domain separation via an intent.
-pub trait SigningIntent {
-    const INTENT: IntentType;
-}
-
-// ---------------------------------
-//          Envelopes
-// ---------------------------------
-
-/// Guardian-signed wrapper - adds timestamp and signature to any data
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
-pub struct GuardianSigned<T> {
-    pub data: T,
-    /// Milliseconds since Unix epoch.
-    pub timestamp_ms: UnixMillis,
-    pub signature: GuardianSignature,
-}
-
-/// Canonical log record written to S3.
-#[derive(Serialize, Deserialize, Debug)]
-pub struct LogRecord {
-    pub session_id: String,
-    pub timestamp_ms: UnixMillis,
-    pub message: LogMessage,
-    /// Present for signed logs; omitted for unsigned logs (currently only OIAttestationUnsigned).
-    pub signature: Option<GuardianSignature>,
-}
-
-/// A verified log record where message authenticity has been checked.
-#[derive(Debug)]
-pub struct VerifiedLogRecord {
-    pub session_id: String,
-    pub timestamp_ms: UnixMillis,
-    pub message: LogMessage,
-}
-
-// ---------------------------------
-//    All requests and responses
-// ---------------------------------
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct SetupNewKeyRequest {
-    key_provisioner_pgp_certs: Vec<PgpPublicCert>,
-    params: SecretSharingParams,
-}
-
-/// `EnclaveSigned<T>`
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
-pub struct SetupNewKeyResponse {
-    pub encrypted_shares: Vec<KPEncryptedShare>,
-    pub share_commitments: ShareCommitments,
-}
 
 /// Operator-supplied bootstrap. A ceremony-mode enclave (setup/rotate) needs only
 /// `s3_config`; a withdraw-mode enclave additionally carries the `WithdrawModeConfig`
@@ -184,72 +92,6 @@ pub struct SetupNewKeyResponse {
 pub struct OperatorInitRequest {
     s3_config: S3Config,
     state: Option<WithdrawModeConfig>,
-}
-
-/// The current KPs' encrypted key shares, assembled into one submission (in the
-/// relay model, by the relay once it has collected enough). Each share's HPKE AAD
-/// binds the enclave's `state_hash` (the `WithdrawModeState` digest), so a share
-/// only decrypts if the KP agreed on the operator-supplied state.
-#[derive(Debug, Clone, PartialEq)]
-pub struct ProvisionerInitRequest {
-    encrypted_shares: Vec<GuardianEncryptedShare>,
-}
-
-/// The withdraw-mode state KPs attest to. Its `digest()` is the `state_hash`:
-/// bound as HPKE AAD on each KP's share and exposed (as a hash) via `GuardianInfo`.
-/// These are exactly the fields whose agreement is enforced *only* via the digest.
-#[derive(Debug, Clone, PartialEq)]
-pub struct WithdrawModeState {
-    /// Current Hashi committee
-    committee: HashiCommittee,
-    /// Limiter config
-    limiter_config: LimiterConfig,
-    /// Limiter state (tokens available, timestamp, seq)
-    limiter_state: LimiterState,
-    /// Raw MPC verifying key (curve point with y-parity preserved). The
-    /// guardian uses this directly for `derive_verifying_key` so the
-    /// 2-of-2 child key in the leaf script matches the MPC signature.
-    hashi_btc_master_pubkey: HashiMasterG,
-}
-
-/// Full operator-supplied withdraw-mode config: the attested `WithdrawModeState`
-/// plus delivery-only fields that are enforced elsewhere and so are excluded from
-/// the digest (the instance via direct share verification; network is share-irrelevant).
-/// Supplied to `operator_init`.
-#[derive(Debug, Clone, PartialEq)]
-pub struct WithdrawModeConfig {
-    state: WithdrawModeState,
-    /// Secret-sharing scheme for the current BTC key (commitments + N + T).
-    secret_sharing_instance: SecretSharingInstance,
-    /// BTC network.
-    network: Network,
-}
-
-/// Setup-mode rotation request, assembled by the operator from the current KPs'
-/// encrypted old shares (each bound to `state.digest()` as AAD) and the shared
-/// rotation target `state`.
-#[derive(Debug, Clone, PartialEq)]
-pub struct RotateKpsRequest {
-    encrypted_old_shares: Vec<GuardianEncryptedShare>,
-    old_instance: SecretSharingInstance,
-    state: RotateKpsState,
-}
-
-/// The shared rotation target all current KPs authorize. Each binds
-/// `state.digest()` as HPKE AAD on its submission, so the enclave only decrypts
-/// ones that agree on it. Old/new (`n`, `t`) may differ.
-#[derive(Debug, Clone, PartialEq, Serialize)]
-pub struct RotateKpsState {
-    /// Armored OpenPGP certs for the new KP set, sorted to a canonical order
-    /// (see `RotateKpsState::new`). Length equals `new_params.num_shares()`.
-    new_kp_pgp_certs: Vec<String>,
-    new_params: SecretSharingParams,
-}
-
-/// `EnclaveSigned<T>`. The new KP set's encrypted shares, returned by `rotate_kps`.
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
-pub struct RotateKpsResponse {
-    pub encrypted_shares: Vec<KPEncryptedShare>,
 }
 
 #[derive(Debug, PartialEq, Clone)]
@@ -292,7 +134,50 @@ pub struct GuardianInfo {
     pub enclave_btc_pubkey: Option<BitcoinPubkey>,
 }
 
-/// An "immediate withdrawal" request. `HashiSigned<T>.`
+// ---------------------------------------
+//    Withdraw mode requests and responses
+// ---------------------------------------
+
+/// Full operator-supplied withdraw-mode config: the attested `WithdrawModeState`
+/// plus delivery-only fields that are enforced elsewhere and so are excluded from
+/// the digest (the instance via direct share verification; network is share-irrelevant).
+/// Supplied to `operator_init`.
+#[derive(Debug, Clone, PartialEq)]
+pub struct WithdrawModeConfig {
+    state: WithdrawModeState,
+    /// Secret-sharing scheme for the current BTC key (commitments + N + T).
+    secret_sharing_instance: SecretSharingInstance,
+    /// BTC network.
+    network: Network,
+}
+
+/// The withdraw-mode state KPs attest to. Its `digest()` is the `state_hash`:
+/// bound as HPKE AAD on each KP's share and exposed (as a hash) via `GuardianInfo`.
+/// These are exactly the fields whose agreement is enforced *only* via the digest.
+#[derive(Debug, Clone, PartialEq)]
+pub struct WithdrawModeState {
+    /// Current Hashi committee
+    committee: HashiCommittee,
+    /// Limiter config
+    limiter_config: LimiterConfig,
+    /// Limiter state (tokens available, timestamp, seq)
+    limiter_state: LimiterState,
+    /// Raw MPC verifying key (curve point with y-parity preserved). The
+    /// guardian uses this directly for `derive_verifying_key` so the
+    /// 2-of-2 child key in the leaf script matches the MPC signature.
+    hashi_btc_master_pubkey: HashiMasterG,
+}
+
+/// The current KPs' encrypted key shares, assembled into one submission (in the
+/// relay model, by the relay once it has collected enough). Each share's HPKE AAD
+/// binds the enclave's `state_hash` (the `WithdrawModeState` digest), so a share
+/// only decrypts if the KP agreed on the operator-supplied state.
+#[derive(Debug, Clone, PartialEq)]
+pub struct ProvisionerInitRequest {
+    encrypted_shares: Vec<GuardianEncryptedShare>,
+}
+
+/// A withdrawal request. `HashiSigned<T>.`
 /// Note: Deserialize is not implemented because UTXOs contain validated addresses.
 /// StandardWithdrawalRequestWire mocks this type with unverified addresses and Deserialize trait.
 #[derive(Debug, Clone, Serialize, PartialEq)]
@@ -316,112 +201,55 @@ pub struct StandardWithdrawalResponse {
 }
 
 /// Committee handoff payload signed by the outgoing committee as
-/// `HashiSigned<CommitteeTransition>`. `new_committee` is the Move BCS
+/// `HashiSigned<CommitteeTransitionRequest>`. `new_committee` is the Move BCS
 /// shape so on-chain and guardian signatures match.
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub struct CommitteeTransition {
+pub struct CommitteeTransitionRequest {
     pub new_committee: crate::move_types::Committee,
 }
 
-// ---------------------------------
-//          Log Messages
-// ---------------------------------
+// ---------------------------------------
+//    Ceremony mode requests and responses
+// ---------------------------------------
 
-/// All log messages emitted by the guardian enclave.
-/// Uses enum discriminator for automatic domain separation between variants.
-#[derive(Debug, Serialize, Deserialize)]
-pub enum LogMessage {
-    Heartbeat { seq: u64 },
-    Init(Box<InitLogMessage>),
-    Withdrawal(Box<WithdrawalLogMessage>),
-    Ceremony(Box<CeremonyLogMessage>),
-    CommitteeUpdate(Box<CommitteeUpdateLogMessage>),
+#[derive(Debug, Clone, PartialEq)]
+pub struct SetupNewKeyRequest {
+    key_provisioner_pgp_certs: Vec<PgpPublicCert>,
+    params: SecretSharingParams,
 }
 
-/// The authoritative share state, written to `ceremony/` after each ceremony.
-/// Carries only instances (commitments + n/t/seq); ciphertexts are delivered
-/// out-of-band and verified against the commitments. A rotation records the
-/// `old_instance` it consumed so the chain is auditable from the log alone.
-#[derive(Debug, Serialize, Deserialize, Clone, PartialEq)]
-pub enum CeremonyLogMessage {
-    /// Initial key setup (`setup_new_key`); `instance` has `sharing_seq` 0.
-    NewKey { instance: SecretSharingInstance },
-    /// Key rotation (`rotate_kps`) from `old_instance` to `new_instance`.
-    Rotate {
-        old_instance: SecretSharingInstance,
-        new_instance: SecretSharingInstance,
-    },
+/// `EnclaveSigned<T>`
+#[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
+pub struct SetupNewKeyResponse {
+    pub encrypted_shares: Vec<KPEncryptedShare>,
+    pub share_commitments: ShareCommitments,
 }
 
-impl CeremonyLogMessage {
-    /// The resulting instance's `sharing_seq` — used as the `ceremony/` object key.
-    pub fn sharing_seq(&self) -> u64 {
-        match self {
-            CeremonyLogMessage::NewKey { instance } => instance.sharing_seq(),
-            CeremonyLogMessage::Rotate { new_instance, .. } => new_instance.sharing_seq(),
-        }
-    }
+/// Ceremony-mode rotation request, assembled by the operator from the current KPs'
+/// encrypted old shares (each bound to `state.digest()` as AAD) and the shared
+/// rotation target `state`.
+#[derive(Debug, Clone, PartialEq)]
+pub struct RotateKpsRequest {
+    encrypted_old_shares: Vec<GuardianEncryptedShare>,
+    old_instance: SecretSharingInstance,
+    state: RotateKpsState,
 }
 
-/// OI: operator_init
-/// PI: provisioner_init
-/// Init messages are expected to be logged in the following order:
-/// OIAttestationUnsigned -> OIGuardianInfo -> PIEnclaveFullyInitialized.
-#[derive(Debug, Serialize, Deserialize)]
-pub enum InitLogMessage {
-    /// Attestation and signing public key posted in /operator_init
-    OIAttestationUnsigned {
-        attestation: Attestation,
-        signing_public_key: GuardianPubKey,
-    },
-    /// Signed GuardianInfo logged in /operator_init (secret-sharing instance,
-    /// state_hash, encryption/BTC pubkeys).
-    OIGuardianInfo(GuardianInfo),
-    /// Threshold reached — enclave fully initialized (happens once). Records the
-    /// ids of the shares that were combined.
-    PIEnclaveFullyInitialized { share_ids: Vec<ShareID> },
+/// The shared rotation target all current KPs authorize. Each binds
+/// `state.digest()` as HPKE AAD on its submission, so the enclave only decrypts
+/// ones that agree on it. Old/new (`n`, `t`) may differ.
+#[derive(Debug, Clone, PartialEq, Serialize)]
+pub struct RotateKpsState {
+    /// Armored OpenPGP certs for the new KP set, sorted to a canonical order
+    /// (see `RotateKpsState::new`). Length equals `new_params.num_shares()`.
+    new_kp_pgp_certs: Vec<String>,
+    new_params: SecretSharingParams,
 }
 
-#[derive(Debug, Serialize, Deserialize)]
-pub enum WithdrawalLogMessage {
-    /// Immediate withdraw success
-    Success {
-        txid: Txid,
-        request_data: StandardWithdrawalRequestWire,
-        request_sign: CommitteeSignature,
-        response: StandardWithdrawalResponse,
-        /// Limiter state after this withdrawal was consumed. The KP rotating in
-        /// the next enclave reads the max-seq Success log and uses its
-        /// `post_state` as the new enclave's initial limiter state.
-        post_state: LimiterState,
-    },
-    /// Immediate withdraw failure
-    Failure {
-        request_data: StandardWithdrawalRequestWire,
-        request_sign: CommitteeSignature,
-        error: GuardianError,
-    },
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub enum CommitteeUpdateLogMessage {
-    /// `from_epoch` is the guardian's current epoch at the time; the
-    /// applied epoch is `new_committee.epoch`. Both are recorded because
-    /// hashi reconfig is sparse — `new_committee.epoch` is not
-    /// necessarily `from_epoch + 1`.
-    Success {
-        from_epoch: u64,
-        new_committee: crate::move_types::Committee,
-        request_sign: CommitteeSignature,
-    },
-    /// `from_epoch` is the guardian's current epoch at the time;
-    /// `new_committee` is what was proposed (and rejected).
-    Failure {
-        from_epoch: u64,
-        new_committee: crate::move_types::Committee,
-        request_sign: CommitteeSignature,
-        error: GuardianError,
-    },
+/// `EnclaveSigned<T>`. The new KP set's encrypted shares, returned by `rotate_kps`.
+#[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
+pub struct RotateKpsResponse {
+    pub encrypted_shares: Vec<KPEncryptedShare>,
 }
 
 // ---------------------------------
@@ -450,26 +278,6 @@ pub struct S3BucketInfo {
 // ---------------------------------
 //          Helper impl's
 // ---------------------------------
-
-impl SigningIntent for LogMessage {
-    const INTENT: IntentType = IntentType::LogMessage;
-}
-
-impl SigningIntent for SetupNewKeyResponse {
-    const INTENT: IntentType = IntentType::SetupNewKeyResponse;
-}
-
-impl SigningIntent for StandardWithdrawalResponse {
-    const INTENT: IntentType = IntentType::StandardWithdrawalResponse;
-}
-
-impl SigningIntent for GuardianInfo {
-    const INTENT: IntentType = IntentType::GuardianInfo;
-}
-
-impl SigningIntent for RotateKpsResponse {
-    const INTENT: IntentType = IntentType::RotateKpsResponse;
-}
 
 impl S3Config {
     pub fn bucket_name(&self) -> &str {
@@ -799,244 +607,6 @@ impl StandardWithdrawalRequest {
     }
 }
 
-impl InitLogMessage {
-    pub const OI_ATTEST_UNSIGNED: &'static str = "oi-attestation-unsigned";
-    pub const OI_GUARDIAN_INFO: &'static str = "oi-guardian-info";
-    pub const PI_FULLY_INITIALIZED: &'static str = "pi-enclave-fully-initialized";
-
-    pub fn log_name(&self, prefix: &str) -> String {
-        let suffix = match self {
-            InitLogMessage::OIAttestationUnsigned { .. } => Self::OI_ATTEST_UNSIGNED.to_string(),
-            InitLogMessage::OIGuardianInfo(_) => Self::OI_GUARDIAN_INFO.to_string(),
-            InitLogMessage::PIEnclaveFullyInitialized { .. } => {
-                Self::PI_FULLY_INITIALIZED.to_string()
-            }
-        };
-
-        format!("{}-{}.json", prefix, suffix)
-    }
-
-    pub fn attestation_object_key(session_id: &str) -> String {
-        format!(
-            "{}/{}-{}.json",
-            S3_DIR_INIT,
-            session_id,
-            Self::OI_ATTEST_UNSIGNED
-        )
-    }
-
-    pub fn guardian_info_object_key(session_id: &str) -> String {
-        format!(
-            "{}/{}-{}.json",
-            S3_DIR_INIT,
-            session_id,
-            Self::OI_GUARDIAN_INFO
-        )
-    }
-}
-
-impl WithdrawalLogMessage {
-    /// Success keys lead with `success-{seq:020}` so that lexicographic listing
-    /// within an hour bucket is also seq-sorted — the last key is the max-seq
-    /// log, which the KP reads to recover limiter state. Failures don't have a
-    /// meaningful seq (the request's seq may be stale), so they use a random
-    /// suffix for dedup.
-    pub fn log_name(&self, prefix: &str) -> String {
-        match self {
-            WithdrawalLogMessage::Success { request_data, .. } => format!(
-                "success-{:020}-{}-wid{}.json",
-                request_data.seq,
-                prefix,
-                self.wid()
-            ),
-            WithdrawalLogMessage::Failure { .. } => {
-                let random_suffix = rand::random::<u32>();
-                format!(
-                    "failure-{}-wid{}-{:08x}.json",
-                    prefix,
-                    self.wid(),
-                    random_suffix
-                )
-            }
-        }
-    }
-}
-
-impl CommitteeUpdateLogMessage {
-    /// Success keys lead with the new epoch (zero-padded) so a lex listing
-    /// is epoch-sorted; failures lead with `failure-` so they sort after
-    /// all successes, leaving the lex-last success key as the latest
-    /// successfully-applied epoch.
-    pub fn log_name(&self, prefix: &str) -> String {
-        match self {
-            CommitteeUpdateLogMessage::Success { new_committee, .. } => {
-                format!("{:020}-{}.json", new_committee.epoch, prefix)
-            }
-            CommitteeUpdateLogMessage::Failure { new_committee, .. } => {
-                let random_suffix = rand::random::<u32>();
-                format!(
-                    "failure-{:020}-{}-{:08x}.json",
-                    new_committee.epoch, prefix, random_suffix
-                )
-            }
-        }
-    }
-}
-
-impl LogMessage {
-    pub fn is_allowed_unsigned(&self) -> bool {
-        if let LogMessage::Init(init_message) = self {
-            matches!(**init_message, InitLogMessage::OIAttestationUnsigned { .. })
-        } else {
-            false
-        }
-    }
-
-    pub fn must_be_signed(&self) -> bool {
-        !self.is_allowed_unsigned()
-    }
-
-    /// The directory under which logs are written. Ends with a slash.
-    pub fn log_dir(&self, timestamp_ms: UnixMillis) -> String {
-        match self {
-            LogMessage::Init(_) => format!("{}/", S3_DIR_INIT),
-            LogMessage::Heartbeat { .. } => {
-                S3HourScopedDirectory::new(S3_DIR_HEARTBEAT, unix_millis_to_seconds(timestamp_ms))
-                    .to_string()
-            }
-            LogMessage::Withdrawal(..) => {
-                S3HourScopedDirectory::new(S3_DIR_WITHDRAW, unix_millis_to_seconds(timestamp_ms))
-                    .to_string()
-            }
-            LogMessage::Ceremony(..) => format!("{}/", S3_DIR_CEREMONY),
-            LogMessage::CommitteeUpdate(..) => format!("{}/", S3_DIR_COMMITTEE_UPDATE),
-        }
-    }
-
-    /// The name of the log.
-    pub fn log_name(&self, prefix: &str) -> String {
-        match self {
-            LogMessage::Init(init_message) => init_message.log_name(prefix),
-            LogMessage::Heartbeat { seq } => format!("{}-{:020}.json", prefix, seq),
-            LogMessage::Withdrawal(withdrawal_message) => withdrawal_message.log_name(prefix),
-            LogMessage::Ceremony(ss) => format!("{:020}-{}.json", ss.sharing_seq(), prefix),
-            LogMessage::CommitteeUpdate(committee_message) => committee_message.log_name(prefix),
-        }
-    }
-
-    pub fn into_init_log(self) -> Option<InitLogMessage> {
-        match self {
-            LogMessage::Init(init_message) => Some(*init_message),
-            _ => None,
-        }
-    }
-}
-
-impl LogRecord {
-    pub fn new(session_id: String, message: LogMessage, signing_key: &GuardianSignKeyPair) -> Self {
-        let timestamp_ms = now_timestamp_ms();
-        if message.is_allowed_unsigned() {
-            Self::unsigned(session_id, message, timestamp_ms)
-        } else {
-            Self::signed(session_id, message, signing_key, timestamp_ms)
-        }
-    }
-
-    /// Full S3 object key for this log record (directory + file name). See the
-    /// `hashi-guardian` README for the canonical per-log-type key layout.
-    pub fn object_key(&self) -> String {
-        let dir = self.message.log_dir(self.timestamp_ms);
-        let log_name = self.message.log_name(&self.session_id);
-        format!("{}{}", dir, log_name)
-    }
-
-    pub fn object_lock_duration(&self) -> Duration {
-        match &self.message {
-            LogMessage::Init(..) => S3_OBJECT_LOCK_DURATION_INIT,
-            LogMessage::Heartbeat { .. } => S3_OBJECT_LOCK_DURATION_HEARTBEAT,
-            LogMessage::Withdrawal(..) => S3_OBJECT_LOCK_DURATION_WITHDRAW,
-            LogMessage::Ceremony(..) => S3_OBJECT_LOCK_DURATION_CEREMONY,
-            LogMessage::CommitteeUpdate(..) => S3_OBJECT_LOCK_DURATION_COMMITTEE_UPDATE,
-        }
-    }
-
-    fn signed(
-        session_id: String,
-        message: LogMessage,
-        signing_key: &GuardianSignKeyPair,
-        timestamp_ms: UnixMillis,
-    ) -> Self {
-        let signed = GuardianSigned::new(message, signing_key, timestamp_ms);
-        Self {
-            session_id,
-            timestamp_ms: signed.timestamp_ms,
-            message: signed.data,
-            signature: Some(signed.signature),
-        }
-    }
-
-    fn unsigned(session_id: String, message: LogMessage, timestamp_ms: UnixMillis) -> Self {
-        assert!(
-            message.is_allowed_unsigned(),
-            "message must be Init(OIAttestationUnsigned)"
-        );
-        Self {
-            session_id,
-            timestamp_ms,
-            message,
-            signature: None,
-        }
-    }
-
-    pub fn verify(self, pub_key: &GuardianPubKey) -> GuardianResult<VerifiedLogRecord> {
-        let session_id = self.session_id;
-        let timestamp_ms = self.timestamp_ms;
-        let message = self.message;
-
-        let message = if message.is_allowed_unsigned() {
-            message
-        } else {
-            let signature = self
-                .signature
-                .ok_or_else(|| InvalidInputs("missing log signature".into()))?;
-            GuardianSigned {
-                data: message,
-                timestamp_ms,
-                signature,
-            }
-            .verify(pub_key)?
-        };
-
-        Ok(VerifiedLogRecord {
-            session_id,
-            timestamp_ms,
-            message,
-        })
-    }
-
-    pub fn verify_unsigned(self) -> GuardianResult<VerifiedLogRecord> {
-        if !self.message.is_allowed_unsigned() {
-            return Err(InvalidInputs(
-                "expected unsigned log record but message requires a signature".into(),
-            ));
-        }
-        Ok(VerifiedLogRecord {
-            session_id: self.session_id,
-            timestamp_ms: self.timestamp_ms,
-            message: self.message,
-        })
-    }
-}
-
-impl WithdrawalLogMessage {
-    pub fn wid(&self) -> WithdrawalID {
-        match self {
-            WithdrawalLogMessage::Success { request_data, .. } => request_data.wid,
-            WithdrawalLogMessage::Failure { request_data, .. } => request_data.wid,
-        }
-    }
-}
-
 pub fn verify_enclave_attestation(_attestation: Attestation) -> GuardianResult<()> {
     // TODO: Implement me
     Ok(())
@@ -1152,11 +722,6 @@ impl From<&WithdrawModeState> for WithdrawModeStateRepr {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use bitcoin::hashes::Hash as _;
-
-    fn set_timestamp(log: &mut LogRecord, timestamp_ms: UnixMillis) {
-        log.timestamp_ms = timestamp_ms;
-    }
 
     #[test]
     fn rotate_kps_state_new_rejects_wrong_cert_count() {
@@ -1187,80 +752,5 @@ mod tests {
         // Same set, different input order ⇒ identical canonical form and digest.
         assert_eq!(a.new_kp_pgp_certs(), b.new_kp_pgp_certs());
         assert_eq!(a.digest(), b.digest());
-    }
-
-    #[test]
-    fn object_key_for_init_attestation_unsigned() {
-        let session_id = "session-a".to_string();
-        let signing_key = GuardianSignKeyPair::from([7u8; 32]);
-        let mut log = LogRecord::new(
-            session_id.clone(),
-            LogMessage::Init(Box::new(InitLogMessage::OIAttestationUnsigned {
-                attestation: vec![1, 2, 3],
-                signing_public_key: signing_key.verification_key(),
-            })),
-            &signing_key,
-        );
-        set_timestamp(&mut log, 1_700_000_000_000);
-
-        assert_eq!(
-            log.object_key(),
-            "init/session-a-oi-attestation-unsigned.json"
-        );
-    }
-
-    #[test]
-    fn object_key_for_heartbeat() {
-        let session_id = "session-b".to_string();
-        let signing_key = GuardianSignKeyPair::from([8u8; 32]);
-        let seq = 42_u64;
-        let timestamp_ms = 1_700_000_000_000;
-
-        let mut log = LogRecord::new(
-            session_id.clone(),
-            LogMessage::Heartbeat { seq },
-            &signing_key,
-        );
-        set_timestamp(&mut log, timestamp_ms);
-
-        assert_eq!(
-            log.object_key(),
-            "heartbeat/2023/11/14/22/session-b-00000000000000000042.json"
-        );
-    }
-
-    #[test]
-    fn object_key_for_withdrawal_success() {
-        let session_id = "session-c".to_string();
-        let signing_key = GuardianSignKeyPair::from([9u8; 32]);
-        let timestamp_ms = 1_700_000_000_000;
-        let wid = WithdrawalID::new([0xcd; 32]);
-        let signed_request =
-            StandardWithdrawalRequest::mock_signed_for_testing_with_wid(Network::Regtest, wid);
-        let (request_sign, request_data) = signed_request.into_parts();
-        let request_data: StandardWithdrawalRequestWire = request_data.into();
-        let seq = request_data.seq;
-
-        let mut log = LogRecord::new(
-            session_id.clone(),
-            LogMessage::Withdrawal(Box::new(WithdrawalLogMessage::Success {
-                txid: Txid::from_slice(&[3u8; 32]).expect("valid txid"),
-                request_data,
-                request_sign,
-                response: GuardianSigned::<StandardWithdrawalResponse>::mock_for_testing().data,
-                post_state: LimiterState {
-                    num_tokens_available: 0,
-                    last_updated_at: 0,
-                    next_seq: seq + 1,
-                },
-            })),
-            &signing_key,
-        );
-        set_timestamp(&mut log, timestamp_ms);
-
-        assert_eq!(
-            log.object_key(),
-            format!("withdraw/2023/11/14/22/success-{seq:020}-session-c-wid{wid}.json"),
-        );
     }
 }
